Shawon

+++ database/migrations/2020_03_21_133631_create_bed_allotments_table.php
@@ -15,8 +15,8 @@ class CreateBedAllotmentsTable extends Migration
     {
         Schema::create('bed_allotments', function (Blueprint $table) {
             $table->bigIncrements('bed_allot_id');
-            $table->integer('bed_allot_bed_id');
-            $table->integer('bed_allot_p_id');
+            $table->unsignedBigInteger('bed_allot_bed_id');
+            $table->unsignedBigInteger('bed_allot_p_id');
             $table->date('bed_allot_start');
             $table->date('bed_allot_discharge');
             $table->integer('patients_p_id');
